Moved role type to package object in state package
