+++ src/main/scala/mesosphere/marathon/core/appinfo/EnrichedTask.scala
@@ -5,7 +5,7 @@ import mesosphere.marathon.core.task.{Task, Tasks}
 import mesosphere.marathon.core.health.Health
 import mesosphere.marathon.core.instance.Instance.AgentInfo
 import mesosphere.marathon.core.instance.{Instance, Reservation}
-import mesosphere.marathon.state.Group.Role
+import mesosphere.marathon.state.Role
 import mesosphere.marathon.state.PathId
 
 case class EnrichedTask(
