+++ src/main/scala/mesosphere/marathon/state/Group.scala
@@ -127,7 +127,6 @@ class Group(
 
 object Group extends StrictLogging {
   type GroupKey = PathId
-  type Role = String
 
   def apply(
     id: PathId,
