+++ src/main/scala/mesosphere/marathon/state/ResourceRole.scala
@@ -4,7 +4,6 @@ package state
 import com.wix.accord._
 import com.wix.accord.dsl._
 import mesosphere.marathon.api.v2.Validation._
-import mesosphere.marathon.state.Group.Role
 
 import scala.collection.SortedSet
