+++ src/main/scala/mesosphere/marathon/state/package.scala
@@ -1,0 +1,7 @@
+package mesosphere.marathon
+
+package object state {
+
+  type Role = String
+
+}
